V2_reasoning_model/requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
openai
langchain
langchain_community
langchain_openai
qdrant-client
chardet 